requirements.txt
psycopg2-binary
bcrypt
cryptography
pycryptodome

